# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "马森雨个人技术博客（masenyu.top）RESTful API 接口文档，涵盖文章、分类、标签、项目、友链、评论、AI助手、数据看板与站点管理等模块。",
        "title": "马森雨个人技术博客 API",
        "contact": {
            "name": "马森雨",
            "url": "https://masenyu.top"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/admin/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "聚合统计文章数、评论数、浏览量、最近趋势与分析数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仪表盘与统计分析"
                ],
                "summary": "获取管理后台仪表盘概览",
                "responses": {
                    "200": {
                        "description": "仪表盘数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.DashboardDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/trends": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CsrfToken": []
                    }
                ],
                "description": "查询指定天数（7~60天，默认14天）的每日PV、UV、评论与新增访客趋势",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仪表盘与统计分析"
                ],
                "summary": "获取访问趋势数据",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 14,
                        "description": "统计天数(7-60)",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "趋势数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.TrendsDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AnalyticsDTO": {
            "type": "object",
            "properties": {
                "failedRequests": {
                    "type": "integer"
                },
                "recentBans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.IPBanDTO"
                    }
                },
                "todayRequests": {
                    "type": "integer"
                },
                "topIPs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.TopIPDTO"
                    }
                },
                "topPaths": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.TopPathDTO"
                    }
                },
                "totalRequests": {
                    "type": "integer"
                },
                "uniqueIPs": {
                    "type": "integer"
                }
            }
        },
...
        "handler.CommentAuthorDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
        "handler.CommentDTO": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/handler.CommentAuthorDTO"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "postId": {
                    "type": "integer"
                },
                "postTitle": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handler.DashboardDTO": {
            "type": "object",
            "properties": {
                "analytics": {
                    "$ref": "#/definitions/handler.AnalyticsDTO"
                },
                "recentComments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.CommentDTO"
                    }
                },
                "stats": {
                    "$ref": "#/definitions/handler.DashboardStatsDTO"
                },
                "trends": {
                    "$ref": "#/definitions/handler.TrendsDTO"
                }
            }
        },
        "handler.DashboardStatsDTO": {
            "type": "object",
            "properties": {
                "approvedCommentCount": {
                    "type": "integer"
                },
                "commentCount": {
                    "type": "integer"
                },
                "draftPostCount": {
                    "type": "integer"
                },
                "hiddenCommentCount": {
                    "type": "integer"
                },
                "postCount": {
                    "type": "integer"
                },
                "publishedPostCount": {
                    "type": "integer"
                },
                "totalViews": {
                    "type": "integer"
                },
                "visitorCount": {
                    "type": "integer"
                }
            }
        },
...
        "handler.IPBanDTO": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "handler.TopIPDTO": {
            "type": "object",
            "properties": {
                "banned": {
                    "type": "boolean"
                },
                "failures": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "lastSeen": {
                    "type": "string"
                },
                "requests": {
                    "type": "integer"
                }
            }
        },
        "handler.TopPathDTO": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "requests": {
                    "type": "integer"
                }
            }
        },
        "handler.TrendPointDTO": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "newVisitors": {
                    "type": "integer"
                },
                "requests": {
                    "type": "integer"
                },
                "uniqueIPs": {
                    "type": "integer"
                }
            }
        },
        "handler.TrendsDTO": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "integer"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.TrendPointDTO"
                    }
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "参数错误"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "请输入 JWT Token，格式为: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CsrfToken": {
            "description": "管理端防跨站请求伪造令牌",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}